% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_value_external.R
\name{compute_economic_value}
\alias{compute_economic_value}
\title{General function to compute an economic value for carcass fat, carcass conformation or carcass weight.}
\usage{
compute_economic_value(pvec_class_freq = NULL, pvec_threshold = NULL,
  pvec_price, pn_mean, pn_sd, pn_delta_mean)
}
\arguments{
\item{pvec_class_freq}{discrete empirical distribution of population for given trait over different classes}

\item{pvec_threshold}{thresholds under continuous distribution given by pricing system}

\item{pvec_price}{vector of prices for given classes}

\item{pn_mean}{current population phenotypic mean}

\item{pn_sd}{current population phenotypic standard deviation}

\item{pn_delta_mean}{small change of current phenotypic population mean}
}
\value{
economic value
}
\description{
This function creates an economic value for either carcass fat and carcass conformation or carcass weight. It uses the "internal finction using thresholds". First it distinguishes between carcass conformation/carcass fat and carcass weight. Carcass weight has thresholds as input parameter, but carcass conformation and carcass fat have population frequencies per class as input parameters. There the thresholds are computed first from the frequencies using the characteristics of normal distributions.
}
