% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_gain_overall_CF.R
\name{compute_genetic_gain_overall_CF}
\alias{compute_genetic_gain_overall_CF}
\title{Genetic Gain in Sfr. / year for when selecting for Carcass fat.}
\usage{
compute_genetic_gain_overall_CF(genetic_var_cov, residual_var_cov,
  offspring, proportion_calves, proportion_adults, male_proportionselected,
  female_proportionselected, male_generationintervall,
  female_generationintervall, economic_weights)
}
\arguments{
\item{genetic_var_cov}{The genetic variance-covariance matrix between the involved traits in the aggregate genotype.}

\item{residual_var_cov}{The genetic variance-covariance matrix between the involved traits in the aggregate genotype.}

\item{proportion_calves}{Proportion of calves in the offspring.}

\item{proportion_adults}{Proportion of adults in the offspring.}

\item{male_proportionselected}{Proportion selected of the offspring of a male selection candidate}

\item{female_proportionselected}{Proportion selected of the offspring of a female selection candidate}

\item{male_generationintervall}{Generation intervall of the male selection path.}

\item{female_generationintervall}{Generation intervall of the female selection path.}

\item{economic_weights}{A vector of economic weights with lenght of number of traits.}

\item{male_offspring}{Number of the assumed offspring of the male selection candidate when selection occurs.}

\item{female_offspring}{Number of the assumed offspring of the female selection candidate when selection occurs.}
}
\value{
Genetic Gain in Sfr. per year
}
\description{
This function computes the genetic gain in Sfr. / year of the selection on carcass fat.
}
