% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_gain_trait.R
\name{compute_genetic_gain_trait}
\alias{compute_genetic_gain_trait}
\title{Correlated genetic gain in trait unit / year for index.}
\usage{
compute_genetic_gain_trait(genetic_var_cov, residual_var_cov, offspring,
  proportion_calves, proportion_adults, male_proportionselected,
  female_proportionselected, male_generationintervall,
  female_generationintervall, economic_weights, trait)
}
\arguments{
\item{genetic_var_cov}{The genetic variance-covariance matrix between the involved traits in the aggregate genotype.}

\item{residual_var_cov}{The genetic variance-covariance matrix between the involved traits in the aggregate genotype.}

\item{proportion_calves}{Proportion of calves in the offspring.}

\item{proportion_adults}{Proportion of adults in the offspring.}

\item{male_proportionselected}{Proportion selected of the offspring of a male selection candidate}

\item{female_proportionselected}{Proportion selected of the offspring of a female selection candidate}

\item{male_generationintervall}{Generation intervall of the male selection path.}

\item{female_generationintervall}{Generation intervall of the female selection path.}

\item{economic_weights}{A vector of economic weights with lenght of number of traits.}

\item{trait}{The name of the trait, of which the genetic gain should be.}

\item{male_offspring}{Number of the assumed offspring of the male selection candidate when selection occurs.}

\item{female_offspring}{Number of the assumed offspring of the female selection candidate when selection occurs.}
}
\value{
Correlated genetic gain in trait unit per year
}
\description{
This function computes the correlated genetic gain in trait unit / year of the selection on the index. In this index all traits are represented in the aggregate genotype.
}
