% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_weight_internal.R
\name{compute_internal_economic_value}
\alias{compute_internal_economic_value}
\title{Internal Function using distribution thresholds.}
\usage{
compute_internal_economic_value(pvec_threshold, pvec_price, pn_mean, pn_sd,
  pn_delta_mean)
}
\arguments{
\item{pvec_threshold}{thresholds under continuous distribution given by pricing system}

\item{pvec_price}{vector of prices for given classes}

\item{pn_mean}{current population phenotypic mean}

\item{pn_sd}{current population phenotypic standard deviation}

\item{pn_delta_mean}{small change of current phenotypic population mean}
}
\value{
economic value
}
\description{
This function creates an economic value for a trait using the distribution thresholds of the traits records over the population. These thresholds are defined by price differences in a specific range of trait values. Price changes per kg carcass weight have been fitted to these thresholds. The difference of price changes between the initial distribution and the distribution with an increased mean divided by the differences in means is the economic value.
}
